requirements.txt
Faker
mysql-connector-python
kafka-python
pyspark